9-gateway/identity/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy
Flask-Migrate
psycopg2
Flask-Script
pyjwt
cryptography
authlib
requests